fix error with sqlite ≥ 3.34
Since sqlite 3.34[1] deleting a record in a fts5 table that doesn't
exist became an error. The resuls it that adding a torrent "corrupts"
the database and magnetico fails with this error:

    Could not add new torrent to the database.
    tx.Exec (INSERT OR REPLACE INTO torrents) database disk image is malformed

The problem can be simply avoided by setting modified_on = discovered_on
when inserting the torrent, instead of updating it afterwards with a
trigger.

[1]: https://github.com/sqlite/sqlite/commit/86f477edaa17767b39c7bae5b67cac8580f7a8c1